requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
h5py
google-generativeai